Create brain_info_store.txt
Take into consideration everything you know about the human brain. Please generate a pseudocode program that reflects a biologically plausible explanation of how information is stored in the human brain. It's OK to speculate, make assumptions here.

Sure, here's an abstract and simplified representation of how information might be stored in the human brain. Please remember that this is an extremely simplified version of a very complex process, and there is much about the brain we still don't fully understand. This model is also a mix of fact and speculation, as you requested, and should be taken as such.

In this pseudocode, I'll focus on two important processes: Encoding (the process of receiving and processing information) and Retrieval (the process of recalling that information). The model will include concepts such as sensory input, attention, perception, memory consolidation, neural plasticity, and association.

Remember that actual human brain doesn't work in this discrete way, and many of the functions outlined here actually take place concurrently, not sequentially. Also, the real process involves many more steps and much more complexity. For example, memories are not stored in a single location, but are rather distributed throughout the brain. Furthermore, our understanding of these processes is constantly evolving. But hopefully, this gives you a rough idea of how some of these processes might be interpreted in a programming context.