Create BrainModel.cpp
Sure, here's a simplified, pseudocode-style C++ class structure that represents the human brain's process of storing information. Let's name the file as BrainModel.cpp.

Please note that this code doesn't include the actual implementation of the methods, as these are largely theoretical and would need much more scientific understanding to implement properly. It's also worth noting that this is a simplification of a complex biological process and is more about structuring code than mimicking the exact processes of the brain.