Create BrainModel.lisp
Take into consideration everything you know about the human brain. Please generate a speculative program written in LISP language that reflects a biologically plausible explanation of how information is stored in the human brain. It's OK to speculate, make assumptions here. Suggest a file name upon completion.

LISP (short for LISt Processing) is one of the oldest high-level programming languages and has a unique syntax. LISP is known for its symbolic processing ability. However, it is also capable of doing numeric and string processing.

The approach here would be to use LISP's list processing abilities to simulate an extremely simplistic model of how neurons might store and process information in the human brain. In reality, the processes involved in the human brain are vastly more complex and intricate than can be represented in a simple LISP program.

Please note that this LISP code is highly speculative and highly simplified, and it's meant more for illustrative purposes than for practical use.

Here's a simple program written in Common Lisp that represents neurons as lists. The program uses functions to represent the addition of synaptic connections and the firing of neurons. We'll name it BrainModel.lisp.

To run this, load the file in your Common Lisp implementation and call (brain-model:main).

Remember, this is a very simplistic and speculative model. The human brain contains billions of interconnected neurons, and the processes through which it stores and processes information are highly complex and not yet fully understood. The above script does not represent these complexities.