Create BrainModel.janus
Take into consideration everything you know about the human brain. Please generate a speculative program written in Janus (time-reversible computing programming language) that reflects a biologically plausible explanation of how information is stored in the human brain. It's OK to speculate, make assumptions here. Suggest a file name upon completion.

Janus is a time-reversible computing language, which is based on a theoretical model of computation where any operation can be undone to return to a previous state. This is a fundamentally different paradigm from conventional programming languages, and it is more commonly used for research into the theoretical underpinnings of computation rather than for modeling complex systems like the human brain.

That being said, it's possible to use the unique features of Janus to create a highly abstract, speculative model of a system with time-reversible characteristics. In the context of the brain, one could speculate that some form of time-reversible process could be involved in the formation and recall of memories, although it's important to note that this is pure speculation and not based on any established neuroscientific theory.

Here is a speculative Janus program which implements a simplistic "memory" in a time-reversible manner. We will call this file BrainModel.janus.

In this script:

forward_memory_store and backward_memory_store are procedures that model the storage and retrieval of a "memory" in a time-reversible manner.
memory_store is a procedure that decides whether to store or retrieve a "memory" based on the input.
This program is meant to represent, in a very abstract and speculative way, the possibility that the processes of memory formation and recall in the brain could have some form of time-reversibility. It should be stressed, however, that this is a very simplistic and abstract model, and it doesn't reflect the current understanding of the highly complex and non-reversible biochemical and electrical processes involved in memory storage and retrieval in the human brain.