Create BrainModelV2.apl
Take into consideration everything you know about the human brain. Please generate a program written in APL programming language that reflects a biologically plausible explanation of how information is stored in the human brain. Focus on two important processes: Encoding (the process of receiving and processing information) and Retrieval (the process of recalling that information). The model will include concepts such as sensory input, attention, perception, memory consolidation, neural plasticity, and association. Suggest a file name upon completion. 

APL is a powerful, concise, and expressive language known for its use of non-standard symbols and its ability to operate on multi-dimensional arrays as a whole. It's not typically used for large-scale simulations or modeling complex biological systems like the human brain, but we can create an abstract model to illustrate some basic concepts.

Let's create an abstract model using the basic principles of encoding and retrieval. This program would involve a data structure to represent the state of the brain (the encoded information), as well as functions to represent the processes of encoding and retrieval.

In this program:

state is a 2-dimensional matrix that represents the encoded information in the brain.
encode is a function that modifies the state based on sensory input and attention. The greater the attention, the stronger the encoding of the input.
retrieve is a function that retrieves information based on a cue. It uses the dot product to compute the association between the cue and the state.
This is a very simplistic model that doesn't take into account many important aspects of how the brain works, such as the temporal dynamics of neural activity, the complex network structure of the brain, or the biochemical processes involved in neural plasticity. However, it does illustrate some basic principles of how information might be stored and retrieved in the brain.

Save this file as BrainModel.apl.

Please note that the above APL code is for illustrative purposes only and may not run as expected, depending on the specific APL interpreter used. It's a pseudo-code representation of a very complex biological system and is highly abstracted. To run this, an APL environment setup would be required and also further refinement of the APL syntax used here.