Create BrainModelV2.hs
Take into consideration everything you know about the human brain. Please generate a program written in Haskell programming language that reflects a biologically plausible explanation of how information is stored in the human brain. Focus on two important processes: Encoding (the process of receiving and processing information) and Retrieval (the process of recalling that information). The model will include concepts such as sensory input, attention, perception, memory consolidation, neural plasticity, and association. Suggest a file name upon completion. 

We can use the Map data structure from the Data.Map module in Haskell to represent the state of the brain, with synapses as the keys and their strengths as the values. This program will also include functions for encoding and retrieval.

In this Haskell program:

Brain is a type alias for Map Int Double, representing a mapping from synapses to their strengths.
emptyBrain is an empty Brain state.
encode is a function that takes a Brain state, a list of sensory inputs (represented as (Int, Double) pairs), and an attention level, and returns a new Brain state where each synapse's strength has been updated based on the corresponding input and the attention level.
retrieve is a function that takes a Brain state and a cue (represented as an Int), and returns the strength of the corresponding synapse (or Nothing if there is no such synapse).
This is a very simplistic model that doesn't take into account many important aspects of how the brain works, such as the temporal dynamics of neural activity, the complex network structure of the brain, or the biochemical processes involved in neural plasticity. However, it does illustrate some basic principles of how information might be stored and retrieved in the brain.

Save this file as BrainModel.hs.

Please note that to run this Haskell program, you will need a Haskell compiler such as GHC (Glasgow Haskell Compiler). Also, remember that this code is highly abstracted and doesn't function as an actual model of the human brain. It's more of a conceptual representation of some of the key processes.